# Conflicts:
#	common-jre/src/test/java/org/geogebra/common/kernel/geos/GeoInputBoxForComplexTest.java
#	common/src/main/java/org/geogebra/common/kernel/arithmetic/variable/VariableReplacerAlgorithm.java

+++ common/src/main/java/org/geogebra/common/euclidian/MyButton.java
@@ -399,8 +399,8 @@ public class MyButton implements Observer {
 	public void setBounds(GRectangle labelRectangle) {
 		x = (int) labelRectangle.getMinX();
 		y = (int) labelRectangle.getMinY();
-		geoButton.setWidth((int) labelRectangle.getWidth());
-		geoButton.setHeight((int) labelRectangle.getHeight());
+		geoButton.setWidth(labelRectangle.getWidth());
+		geoButton.setHeight(labelRectangle.getHeight());
 	}
 
 	/**
